refactor: remove services
