--- src/service/rooms.ts
@@ -1,18 +1,0 @@
-import {RoomEntity} from "../database/types";
-import {addRoom, addToRoom, getEmptyRooms} from "../database/rooms";
-
-export interface AddUserToRoomRequest {
-    indexRoom: number,
-}
-
-export const getAvailableRooms = (): RoomEntity[] => {
-    return getEmptyRooms()
-}
-
-export const createRoom = (userId: number) => {
-    addRoom(userId)
-}
-
-export const addUserToRoom = (userId: number, request: AddUserToRoomRequest): RoomEntity => {
-    return addToRoom(userId, request.indexRoom)
-}
